whoopsie, didn't commit in a while. seems like lots of stuff happened: - GRSEC PATCHES ARE BACK LESSGOOOOOOOO - kernels configs don't work yet but some are there already - sysctl docs and sources added - boot params added, attributed - added some placeholder modules so that my adhd brain doesn't forget what's still to be done - removed apparmor-d as it is being worked on in its own repo - added hide-hardware-info, ly kicksecure :* - changed some misc configs which i can't remember, ggez

+++ run-test.sh
@@ -3,7 +3,7 @@
 # pass any arg to run in headless mode
 extraargs=""
 rmtmprun() { sudo rm ./run.sh; }
-sudo rm nixos.qcow2
+[ -f nixos.qcow2 ] && sudo rm nixos.qcow2
 nix build .#nixosConfigurations.test.config.system.build.vm --show-trace || exit 1
 cp ./result/bin/run-nixos-vm ./run.sh
 trap rmtmprun EXIT
